refactor(utils): extract status cache and user activity
